third_party/WebKit/LayoutTests/editing/execCommand/indent-block-in-list.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<div contenteditable><ul><li><hr>|</li></ul></div>',
  'indent',
  [
    '<div contenteditable><ul>',
      '<li>',
        '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
          '<li><hr>|</li>',
        '</blockquote>',
      '</li>',
    '</ul></div>',
  ]),
  'Indent command should not crash in list with HR');
</script>
